Initial updates to Ark:SE specific settings

+++ scripts/Server-Tools/Config-Tools.ps1
@@ -1,9 +1,8 @@
 $defaultConfigDirectory='/app/configs/'
-$destinationConfigDirectory=(Join-Path '/app/server' '/path/to/config/folder')
+$destinationConfigDirectory=(Join-Path '/app/server' '/ShooterGame/Saved/Config/LinuxServer/')
 $configFiles=@{
   'Game.ini'=$destinationConfigDirectory
-  'Engine.ini'=$destinationConfigDirectory
-  'Server.ini'=$destinationConfigDirectory
+  'GameUserSettings.ini'=$destinationConfigDirectory
 }
 
 function Copy-Configs
@@ -11,10 +10,13 @@ function Copy-Configs
   foreach ($file in $configFiles.GetEnumerator())
   {
     $sourceFile = (Join-Path $defaultConfigDirectory $file.Name)
-    if (-Not (Test-Path $file.Value)) { mkdir -p $($file.Value) } # New-Item does really weird things in here
-    Copy-Item $sourceFile $file.Value
-  }
-}
+    if (Test-Path $sourceFile)
+    {
+      if (-Not (Test-Path $file.Value)) { mkdir -p $($file.Value) } # New-Item does really weird things in here
+      Copy-Item $sourceFile $file.Value
+    } # if (Test-Path $sourceFile)
+  } # foreach ($file in $configFiles.GetEnumerator())
+} # function Copy-Configs
 
 function Update-ConfigFile
 {
